refactor: separate main code into separate files
